# remove_fixed_joints! of Atlas in benchmark